Separated refund test case
